requirements.txt
django==2.1.1
djangorestframework==3.8.2
requests==2.19.1
mock==2.0.0
